+++ i18n/src/main/kotlin/com/kamelia/sprinkler/i18n/Internal.kt
@@ -1,9 +1,12 @@
 @file:PackagePrivate
 package com.kamelia.sprinkler.i18n
 
+import com.fasterxml.jackson.databind.ObjectMapper
 import com.kamelia.sprinkler.util.unsafeCast
 import com.zwendo.restrikt.annotation.PackagePrivate
 import org.intellij.lang.annotations.Language
+import org.yaml.snakeyaml.Yaml
+import java.io.File
 import java.util.*
 
 @Language("RegExp")
@@ -13,61 +16,21 @@ internal val KEY_IDENTIFIER_REGEX = KEY_IDENTIFIER_REGEX_STRING.toRegex()
 
 internal val FULL_KEY_REGEX = """$KEY_IDENTIFIER_REGEX_STRING(?:\.$KEY_IDENTIFIER_REGEX_STRING)*""".toRegex()
 
-internal fun Map<*, *>.prettyPrint(): String = buildString {
-    append("{\n")
-    val queue = ArrayDeque<Triple<Int, String, Map<*, *>>>()
-    queue.addFirst(Triple(0, "", this@prettyPrint))
-    while (queue.isNotEmpty()) {
-        val (indentLevel, name, map) = queue.removeFirst()
-        prettyPrint(map, queue, indentLevel, name)
-    }
-    append("\n}")
+fun jsonParser(): I18nFileParser = I18nFileParser.from { content ->
+    ObjectMapper().readValue(content, HashMap::class.java).unsafeCast()
 }
 
-private fun StringBuilder.prettyPrint(
-    map: Map<*, *>,
-    queue: ArrayDeque<Triple<Int, String, Map<*, *>>>,
-    indentLevel: Int,
-    keyName: String,
-) {
-    val notEmptyKey = keyName.isNotEmpty()
-    if (notEmptyKey) {
-        repeat(indentLevel) { append("    ") }
-        append("\"")
-        append(keyName)
-        append("\": {\n")
-    }
+fun yamlParser(): I18nFileParser = I18nFileParser.from { Yaml().load(it) }
 
-    map.entries.forEachIndexed { index, (key, value) ->
-        if (value is Map<*, *>) {
-            queue.addFirst(Triple(indentLevel + 1, key.toString(), value.unsafeCast()))
-            return@forEachIndexed
-        }
-        repeat(indentLevel + 1) { append("    ") }
-        append("\"")
-        append(key)
-        append("\": \"")
-        append(value)
-        append("\"")
-        if (index < map.size - 1 || queue.peekFirst()?.first == indentLevel + 1) {
-            append(",")
-        }
-        append("\n")
-    }
-
-    if (notEmptyKey) {
-        val nextIndentation = queue.peekFirst()?.first ?: 1
-        val difference = indentLevel - (nextIndentation - 1)
-        val notLast = queue.isNotEmpty()
-        for (i in 0 until difference) {
-            repeat(indentLevel - i) { append("    ") }
-            append("}")
-            if (i == difference - 1 && queue.isNotEmpty()) {
-                append(",")
-            }
-            if (notLast || i < difference - 1) {
-                append("\n")
-            }
-        }
-    }
+fun main() {
+    val translator = Translator.builder(Locale.FRANCE)
+        .addFile(File("translations"), yamlParser())
+        .build()
+    val translator2 = TranslatorBuilder(Locale.FRANCE)
+        .addFile(File("foo.json"), jsonParser()) { Locale.ENGLISH }
+        .build()
+    val str = translator.section("pages.login").prettyDisplay(Locale.ENGLISH)
+    val str2 = translator2.prettyDisplay(Locale.ENGLISH)
+    println(translator.toMap() == translator2.toMap())
+    println(str2)
 }
